django_app/requirements.txt
django>=4.2
whitenoise
requests
anthropic
python-dotenv
gunicorn
